refactor: change ab testing and fix desktop issues
