+++ src/components/page/JoinNow.svelte
@@ -1,240 +1,48 @@
-<script>
-  import { onMount } from "svelte";
-  import Arrow from "@components/images/Arrow.svelte";
-
-  export let infoBlockGroup;
-
-  const infoBlockGroups = {
-    CONTROL: "2023Sep Control",        // Standard join form
-    IMPACT: "2023Sep Impact",  // Join form with image/impact
-  };
-
-  function scrollToNextSection(){
-    let el = document.querySelector('#join-form');
-
-    window.scrollTo({
-      top: el.offsetTop - 100,
-      behavior: 'smooth'
-    })
-  }
-
-  function toggleInput(focus){
-    let input = document.querySelector('#first_name');
-    focus ? input.focus() : input.blur();
-  }
-
-
-  onMount(async () => {
-    // Track the details text display
-    analytics.track("JoinNow ViewedByABTestParticipant", {
-      infoBlockGroup,
-    });
-
-
-    const observer = new IntersectionObserver((entries) =>{
-      entries.forEach((entry) =>{
-
-        if(entry.isIntersecting){
-          scrollToNextSection()
-          toggleInput(true)
-        }
-      })
-    })
-    const animateElements = document.querySelectorAll('.impact__form--container');
-    animateElements.forEach((el) => observer.observe(el));
-
-  });
-</script>
-{#if infoBlockGroup === infoBlockGroups.IMPACT}
-<div class="impact">
-  <div class="impact__image">
-    <a href="/" class="button secondary back-button"><span class="back-button--icon"><Arrow direction="left" colour="white" /></span><span class="back-button--text">Back</span></a>
-    <div class="impact__image--content">
-      <p>Join more</p>
-      <p>than 40,000</p>
-      <p>people</p>
-      <p>investing in</p>
-      <p class="heading--green">climate</p>
-      <p class="heading--green">solutions.</p>
-    </div>
+<form class="join-form" method="GET" action="https://join.futuresuper.com.au/">
+  <h1>Join Future Super</h1>
+  <div class="time-row">
+    <img src="/images/clock2.gif" alt="clock" class="clock" />
+    <h4>Joining takes about 4 minutes.</h4>
   </div>
-
-  <form class="impact__form" method="GET" action="https://join.futuresuper.com.au/">
-    <div  id="join-form" class="impact__form--container">
-      <h2 class="impact__form--heading">Join Future Super</h2>
-      <div class="time-row">
-        <img src="/images/clock2.gif" alt="clock" class="clock" />
-        <h4>Takes about 4 minutes.</h4>
-      </div>
-      <div class="info-block">
-        <p class="info-block__heading">What you'll need</p>
-        <ul class="info-block__list">
-          <li>Your Tax File Number</li>
-          <li>An Australian ID or Medicare card to find your super and complete a transfer*</li>
-        </ul>
-      </div>
-      <p>
-        <label
-        >First Name<input
-                type="text"
-                id="first_name"
-                name="first_name"
-                required
-        /></label
-        >
-      </p>
-      <p>
-        <label>Email ^<input type="email" name="email" required /></label>
-      </p>
-      <input type="text" id="referer" name="ReferCode" style="display:none" />
-      <p>
-        <button type="submit" class="primary">Next →</button>
-      </p>
-      <p class="disclaimer">
-        * Please note that you don't need to transfer funds to create an account
-        with Future Super.<br /><br />
-        ^ By providing your email address, you consent and authorise us to send you communications
-        or information, including information required by law, via email or similar technologies.
-        Your details will never be passed onto a third party other than in accordance
-        with our <a href="/privacy-policy">Privacy Policy</a>. You can elect to
-        receive communications by post at any time by contacting Future Super on
-        1300 658 422 or via email at info@myfuturesuper.com.au or in writing at PO
-        Box 1282, Albury, NSW 2640.
-      </p>
-    </div>
-
-  </form>
-</div>
-{:else}
-  <form class="join-form" method="GET" action="https://join.futuresuper.com.au/">
-    <h1>Join Future Super</h1>
-    <div class="time-row">
-      <img src="/images/clock2.gif" alt="clock" class="clock" />
-      <h4>Joining takes about 4 minutes.</h4>
-    </div>
-    <div class="info-block">
-      <p class="info-block__heading">What you'll need</p>
-      <ul class="info-block__list">
-        <li>Your Tax File Number</li>
-        <li>An Australian ID or Medicare card to find your super and complete a transfer*</li>
-      </ul>
-    </div>
-    <p>
-      <label
-      >First Name<input
-              type="text"
-              id="first_name"
-              name="first_name"
-              required
-      /></label
-      >
-    </p>
-    <p>
-      <label>Email ^<input type="email" name="email" required /></label>
-    </p>
-    <input type="text" id="referer" name="ReferCode" style="display:none" />
-    <p>
-      <button type="submit" class="primary">Next →</button>
-    </p>
-    <p class="disclaimer">
-      * Please note that you don't need to transfer funds to create an account
-      with Future Super.<br /><br />
-      ^ By providing your email address, you consent and authorise us to send you communications
-      or information, including information required by law, via email or similar technologies.
-      Your details will never be passed onto a third party other than in accordance
-      with our <a href="/privacy-policy">Privacy Policy</a>. You can elect to
-      receive communications by post at any time by contacting Future Super on
-      1300 658 422 or via email at info@myfuturesuper.com.au or in writing at PO
-      Box 1282, Albury, NSW 2640.
-    </p>
-  </form>
-{/if}
-
+  <div class="info-block">
+    <p class="info-block__heading">What you'll need</p>
+    <ul class="info-block__list">
+      <li>Your Tax File Number</li>
+      <li>An Australian ID or Medicare card to find your super and complete a transfer*</li>
+    </ul>
+  </div>
+  <p>
+    <label
+    >First Name<input
+            type="text"
+            id="first_name"
+            name="first_name"
+            required
+    /></label
+    >
+  </p>
+  <p>
+    <label>Email ^<input type="email" name="email" required /></label>
+  </p>
+  <input type="text" id="referer" name="ReferCode" style="display:none" />
+  <p>
+    <button type="submit" class="primary">Next →</button>
+  </p>
+  <p class="disclaimer">
+    * Please note that you don't need to transfer funds to create an account
+    with Future Super.<br /><br />
+    ^ By providing your email address, you consent and authorise us to send you communications
+    or information, including information required by law, via email or similar technologies.
+    Your details will never be passed onto a third party other than in accordance
+    with our <a href="/privacy-policy">Privacy Policy</a>. You can elect to
+    receive communications by post at any time by contacting Future Super on
+    1300 658 422 or via email at info@myfuturesuper.com.au or in writing at PO
+    Box 1282, Albury, NSW 2640.
+  </p>
+</form>
 
 <style lang="scss">
   @use "../../styles/" as *;
-
-  .impact{
-    display: flex;
-    @media (max-width: 800px) {
-      flex-direction: column;
-    }
-
-    &__image{
-      background-image: url('/images/join-form-background.png');
-      background-size: cover;
-      background-position: 0;
-      background-repeat: no-repeat;
-      min-height: 100vh;
-      width: 52%;
-
-      display: flex;
-      align-items: stretch;
-      justify-content: end;
-
-
-
-      position: relative;
-
-      @media (max-width: 800px) {
-        width: 100%;
-        border-radius: 0 0 40px 40px;
-        background-position: 35%;
-      }
-
-      &--content{
-        @media (max-width: 800px) {
-          font-size: 32px;
-        }
-        padding-right: 15%;
-        padding-top: 15%;
-        font-size: 64px;
-        color: white;
-        display: flex;
-        flex-direction: column;
-        align-items: end;
-        justify-content: center;
-        p{
-          background-color: $black;
-          padding: 0 5px;
-          &.heading--green{
-            color: $black;
-            background-color: #77FD79;
-          }
-        }
-
-      }
-    }
-
-    &__form{
-      border-radius: 32px 0 0 32px;
-
-      margin-left: -2%;
-      z-index: 10;
-      background-color: white;
-      width: 50%;
-      display: flex;
-      align-items: center;
-      justify-content: center;
-      @media (max-width: 800px) {
-        width: 100%;
-        border-radius: 0;
-        padding: 240px 0;
-      }
-
-      &--container{
-        padding: 40px;
-        max-width: 550px;
-      }
-
-      &--heading{
-        margin-bottom: 8px;
-
-        font-size: calc(30px + 1vw);
-      }
-    }
-  }
-
   .join-form{
     margin: 20px 0 100px 0;
     border: 1px solid $black;
@@ -271,56 +79,6 @@
   p.disclaimer {
     margin-top: 40px;
   }
-
-  .button{
-    &.secondary{
-      background-color: transparent;
-      border-color: white;
-      @media (max-width: 800px) {
-        padding: 0;
-        display: flex;
-        justify-content: center;
-        align-items: center;
-        width: 45px;
-        height: 45px;
-        background-color: #A5A5A5;
-        border-color: $green;
-        border-width: 1px;
-      }
-
-    }
-  }
-
-  .back-button{
-    position: absolute;
-    top: 65px;
-    left: 58px;
-    @media (max-width: 800px) {
-      top: 12px;
-      left: 25px;
-      position: fixed;
-      z-index: 20;
-    }
-
-    &--icon{
-      width: 20px;
-      height: 20px;
-      margin-right: 15px;
-      @media (max-width: 800px) {
-        margin-right: 0;
-        filter: invert(1);
-
-      }
-    }
-
-    &--text{
-      @media (max-width: 800px) {
-        display: none;
-      }
-    }
-  }
-
-
 
   h4 {
     margin-bottom: 0;
